+++ client/src/components/Form.vue
@@ -1353,7 +1353,6 @@
                 return o
               })
             }
-
         }
 
         // console.log("invoking field expressions and queries")