+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   new function fn.fnCidr
+
 ## [4.0.7] - 2023-05-01
 
 ### Fixed